DOCS: [개선] 쿼리 파일로 분리, 폴더 구조 개선
